+++ src/lib/pages/tx-details/components/tx-message/TxMsgExpand.tsx
@@ -41,12 +41,12 @@ export const TxMsgExpand = ({
 
   const { "@type": type, ...body } = msgBody;
   const isIBC =
-    Boolean(log?.events?.find((event) => event.type === "send_packet")) ||
+    Boolean(log.events.find((event) => event.type === "send_packet")) ||
     type.startsWith("/ibc");
-  let msgIcon: IconKeys = "info-circle";
-  let content: ReactNode;
   const isOpinit = Boolean(type.startsWith("/opinit"));
 
+  let msgIcon: IconKeys = "info-circle";
+  let content: ReactNode;
   switch (type) {
     case "/cosmwasm.wasm.v1.MsgStoreCode":
       msgIcon = "upload";