+++ src/lib/utils/tx/__test__/extractTxLogs.test.ts
@@ -1,0 +1,12 @@
+import { extractTxLogs } from "../extractTxLogs";
+
+import { fromLogs, fromEvents } from "./extractTxLogs.example";
+
+describe("extractTxLogs", () => {
+  test("from logs", () => {
+    expect(extractTxLogs(fromLogs.txData)).toEqual(fromLogs.result);
+  });
+  test("from events", () => {
+    expect(extractTxLogs(fromEvents.txData)).toEqual(fromEvents.result);
+  });
+});